docs: Fix documentation for `timeout` parameter (int -> float)

+++ asyncpg/connection.py
@@ -128,7 +128,7 @@ class Connection:
         """Create a *prepared statement* for the specified query.
 
         :param str query: Text of the query to create a prepared statement for.
-        :param int timeout: Optional timeout value in seconds.
+        :param float timeout: Optional timeout value in seconds.
 
         :return: A :class:`~prepared_stmt.PreparedStatement` instance.
         """
@@ -140,7 +140,7 @@ class Connection:
 
         :param str query: Query text.
         :param args: Query arguments.
-        :param int timeout: Optional timeout value in seconds.
+        :param float timeout: Optional timeout value in seconds.
 
         :return list: A list of :class:`Record` instances.
         """
@@ -156,7 +156,7 @@ class Connection:
         :param args: Query arguments.
         :param int column: Numeric index within the record of the value to
                            return (defaults to 0).
-        :param int timeout: Optional timeout value in seconds.
+        :param float timeout: Optional timeout value in seconds.
                             If not specified, defaults to the value of
                             ``command_timeout`` argument to the ``Connection``
                             instance constructor.
@@ -175,7 +175,7 @@ class Connection:
 
         :param str query: Query text
         :param args: Query arguments
-        :param int timeout: Optional timeout value in seconds.
+        :param float timeout: Optional timeout value in seconds.
 
         :return: The first row as a :class:`Record` instance.
         """
